# api/api-spec.yaml
openapi: 3.0.0
servers:
  - description: API on Test Server
    url: https://csse-s302g4.canterbury.ac.nz/test/api/
  - description: API on Production Server
    url: https://csse-s302g4.canterbury.ac.nz/prod/api/
info:
  description: API specification for the SENG302 2021 project
  version: 6.2.4
  title: REUSEABILITY
  contact:
    email: s302team400@cosc.canterbury.ac.nz
  license:
    name: MIT License
    url: 'https://opensource.org/licenses/MIT'
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  '/images':
    post:
      tags:
        - 'U16: Product Images'
        - 'Catalogue / Products'
      summary: Upload a new image for some photo
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: uncheckedImageType
          in: query
          required: true
          schema:
            type: string
          description: >-
            which entity will use this image

            rules (enum):

            * USER

            * BUSINESS

            * PRODUCT
        - name: userID
          in: query
          required: false
          schema:
            type: integer
          description: ID of use this image related to
        - name: businessID
          in: query
          required: false
          schema:
            type: integer
          description: ID of business this image related to
        - name: productID
          in: query
          required: false
          schema:
            type: string
          description: ID of product this image related to
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                images:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Image created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  imageId:
                    type: integer
                    description: The ID of the image that has just been created
                    example: 1234
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden:
            Returned when a user tries to add an image for another user AND user is not a global application admin
            OR when a user tries to add an image for a business that he do not administer AND the user is not a global application admin
            OR when a user tries to add an image for a product that it is in a catalogue for a business they do not administer AND the user is not a global application admin'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
  '/images/{id}/makePrimary':
    put:
      tags:
        - 'U16: Product Images'
        - 'Catalogue / Products'
      summary: Set an existing image to be the new primary image
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: ID of the image to set as being the primary image for product
          required: true
          schema:
            type: integer
        - name: uncheckedImageType
          in: query
          required: true
          schema:
            type: string
          description: >-
            which entity will use this image

              rules (enum):

              * USER

              * BUSINESS

              * PRODUCT
        - name: userID
          in: query
          required: false
          schema:
            type: integer
          description: ID of use this image related to
        - name: businessId
          in: query
          description: ID of business that owns product
          required: true
          schema:
            type: integer
        - name: productId
          in: query
          description: ID (Product Code) of product to update primary image for
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Primary image successfully updated
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden:
           Returned when a user tries to update the image primary for another user AND user is not a global application admin
           OR when a user tries to update the image primary for a business that he do not administer AND the user is not a global application admin
           OR when a user tries to update the image primary for a product that it is in a catalogue for a business they do not administer AND the user is not a global application admin'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequestError:
      description: There was some error with the data supplied by the user, appropriate error message(s) should be shown to user
    UnauthorizedError:
      description: Access token is missing or invalid
...
    NotAcceptableError:
      description: 'The requested route does exist (so not a 404) but some part of the request is not acceptable, for example trying to access a resource by an ID that does not exist.'
...